IMU localizer ready to test
.
